requirements.txt
requests
python-dotenv
pypdf
numpy
torch; sys_platform == 'linux' or sys_platform == 'darwin'
torchvision; sys_platform == 'linux' or sys_platform == 'darwin'
